feat: 멤머 클래스 예외처리 핸들러 추가
